Better Unicode support in status-column alignment.
`buildTestOutput` aligns the column of [OK] or [FAIL] messages neatly along
a vertical line, to ensure readable output. It does so by measuring the
width of the name-strings, as of now, using the simple `Prelude.length`
function.

But even with a monospaced font, Unicode text generally can't be relied upon to
have a width proportional to the count of code points. Notable, CJK characters
usually need twice the width of Western characters. As a result, Tasty may
produce output like this:

    𝑎 + 𝑏 + 𝑐 + 𝑑  &~?  ㄜ+ㄑ :=: ㄑ+ㄜ:            OK
    𝑎 + 𝑐 + 𝑏  &~?  𝑎+𝑐 :=: ξ:                  OK
    𝑎 + 𝑏 + 𝑐  &~?  𝑏+𝑐 :=: 𝑐+𝑏  &~? 𝑎+𝑐 :=: ξ: OK
    𝑎*𝑥 + 𝑏*𝑥 + 𝑐  &~: ㄏ*ㄘ+ㄐ*ㄘ :=: (ㄏ+ㄐ)*ㄘ:     OK
    (𝑎+𝑏)*𝑥 + 𝑐  &~: (ㄏ+ㄐ)*ㄘ :=: ㄏ*ㄘ+ㄐ*ㄘ:       OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑎*𝑏 :=: 𝑏*𝑎:                    OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑏*𝑐 :=: 𝑐*𝑏:                    OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑐*𝑑 :=: 𝑑*𝑐:                    OK
    𝑎 + 𝑏 - 𝑐 &~: 𝑏-𝑐 :=: (-𝑐)+𝑏:               FAIL

At least under Unix, the `wcwidth` function can be used to learn, for most of
Unicode, how wide the individual `Char` code points will come out on the
terminal, thus we can more reliably estimate the width and improve the
above to

    𝑎 + 𝑏 + 𝑐 + 𝑑  &~?  ㄜ+ㄑ :=: ㄑ+ㄜ:           OK
    𝑎 + 𝑐 + 𝑏  &~?  𝑎+𝑐 :=: ξ:                     OK
    𝑎 + 𝑏 + 𝑐  &~?  𝑏+𝑐 :=: 𝑐+𝑏  &~? 𝑎+𝑐 :=: ξ:    OK
    𝑎*𝑥 + 𝑏*𝑥 + 𝑐  &~: ㄏ*ㄘ+ㄐ*ㄘ :=: (ㄏ+ㄐ)*ㄘ: OK
    (𝑎+𝑏)*𝑥 + 𝑐  &~: (ㄏ+ㄐ)*ㄘ :=: ㄏ*ㄘ+ㄐ*ㄘ:   OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑎*𝑏 :=: 𝑏*𝑎:                       OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑏*𝑐 :=: 𝑐*𝑏:                       OK
    𝑎*𝑏*𝑐*𝑑 &~: 𝑐*𝑑 :=: 𝑑*𝑐:                       OK
    𝑎 + 𝑏 - 𝑐 &~: 𝑏-𝑐 :=: (-𝑐)+𝑏:                  FAIL