Add showDefaultValue to IsOption for default argument rendering in --help
This introduces a new `showDefaultValue` method to `IsOption` for the
purpose of making it possible to display `defaultValue`s in `--help`
output declaratively. For various reasons, `optparse-applicative`'s
own `showDefaultValue` combinator is unsuitable for this purpose
(see the discussion in #269), so we do the next best thing by rolling
our own way to do it. Various options in `tasty` itself can make use
of `showDefaultValue`, including `AnsiTricks`, `UseColor`, and
`NumThreads`.

This approach isn't perfect, as there are various ways you can shoot
yourself in the foot if you construct a `Parser` recklessly. For this
reason, this patch also adds various validity checks that emit
warnings if you do silly things. See
`Test.Tasy.CmdLine.finalizeCLParser`.

In order to propagate the warnings messages from `finalizeCLParser` to
`IO`, this patch also changes the return types of `optionParser` and
`suiteOptionParser` to include a `[String]` alonside the accompanying
`Parser`. Since these functions are exported, this is a breaking API
change.

Fixes #269.