launchTestTree: use IntMap strict in values
At the moment we use `Data.IntMap` which is lazy in values,
and we initialise values with thunks `testStatus <$> testActions`.
This is outright silly: `testStatus` is just a selector for `TVar`
field, forcing it to WHNF can do no harm.

We also take an opportunity to replace `IntMap.fromList` with
`IntMap.fromDistinctAscList`.